+++ lib/src/models/rules/insert.dart
@@ -2,6 +2,7 @@ import 'package:tuple/tuple.dart';
 
 import '../../models/documents/document.dart';
 import '../documents/attribute.dart';
+import '../documents/nodes/embeddable.dart';
 import '../documents/style.dart';
 import '../quill_delta.dart';
 import 'rule.dart';
@@ -239,6 +240,7 @@ class ResetLineFormatOnNewLineRule extends InsertRule {
 }
 
 /// Handles all format operations which manipulate embeds.
+/// This rule wraps line breaks around video, not image.
 class InsertEmbedsRule extends InsertRule {
   const InsertEmbedsRule();
 
@@ -246,6 +248,12 @@ class InsertEmbedsRule extends InsertRule {
   Delta? applyRule(Delta document, int index,
       {int? len, Object? data, Attribute? attribute}) {
     if (data is String) {
+      return null;
+    }
+
+    assert(data is Map);
+
+    if (!(data as Map).containsKey(BlockEmbed.videoType)) {
       return null;
     }