+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [3.9.6]
+* Apply locale to QuillEditor(contents).
+
+# [3.9.5]
+* Fix image pasting.
+
 # [3.9.4]
 * Hiding dialog after selecting action for image.